requirements.txt
python-telegram-bot[webhooks]
openai
google-genai
pillow